Cleanup & Prep
CodeMaid cleanup. Create Household works; Create bank account works.

+++ FinancialTracker_Web/Controllers/HouseholdsController.cs
@@ -1,104 +1,88 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
+﻿using FinancialTracker_Web.Models;
+using Microsoft.AspNet.Identity;
+using System;
 using System.Data.Entity;
 using System.Linq;
 using System.Net;
-using System.Web;
 using System.Web.Mvc;
-using FinancialTracker_Web.Models;
 
 namespace FinancialTracker_Web.Controllers
 {
     public class HouseholdsController : Controller
     {
-        private ApplicationDbContext db = new ApplicationDbContext();
+        private AppDbContext db = new AppDbContext();
 
         // GET: Households
-        public ActionResult Index()
-        {
+        public ActionResult Index() {
             return View(db.Households.ToList());
         }
 
         // GET: Households/Details/5
-        public ActionResult Details(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+        public ActionResult Details() {
+            var house = ApplicationUser.GetFromDb(User, db).Household;
+            if( house == null ) {
+                return RedirectToAction("Index", "Home");
             }
-            Household household = db.Households.Find(id);
-            if (household == null)
-            {
-                return HttpNotFound();
-            }
-            return View(household);
+            return View(house);
         }
 
         // GET: Households/Create
-        public ActionResult Create()
-        {
+        public ActionResult Create() {
             return View();
         }
 
         // POST: Households/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Name,Greeting")] Household household)
-        {
-            if (ModelState.IsValid)
-            {
+        public ActionResult Create([Bind(Include = "Name,Greeting")] Household household) {
+            if( ModelState.IsValid ) {
+                household.Members.Add(ApplicationUser.GetFromDb(User, db));
+                household.CreatorId = User.Identity.GetUserId();
+                household.CreatedAt = DateTime.Now;
+
                 db.Households.Add(household);
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Index", "Home");
             }
 
             return View(household);
         }
 
         // GET: Households/Edit/5
-        public ActionResult Edit(int? id)
-        {
-            if (id == null)
-            {
+        public ActionResult Edit(int? id) {
+            if( id == null ) {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             Household household = db.Households.Find(id);
-            if (household == null)
-            {
+            if( household == null ) {
                 return HttpNotFound();
             }
             return View(household);
         }
 
         // POST: Households/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Name,Greeting")] Household household)
-        {
-            if (ModelState.IsValid)
-            {
+        public ActionResult Edit([Bind(Include = "Id,AccountName,Greeting")] Household household) {
+            if( ModelState.IsValid ) {
                 db.Entry(household).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-            return View(household);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Households/Delete/5
-        public ActionResult Delete(int? id)
-        {
-            if (id == null)
-            {
+        public ActionResult Delete(int? id) {
+            if( id == null ) {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             Household household = db.Households.Find(id);
-            if (household == null)
-            {
+            if( household == null ) {
                 return HttpNotFound();
             }
             return View(household);
@@ -107,18 +91,19 @@ namespace FinancialTracker_Web.Controllers
         // POST: Households/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
-        {
+        public ActionResult DeleteConfirmed(int id) {
             Household household = db.Households.Find(id);
-            db.Households.Remove(household);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if( household != null ) {
+                household.Members.Clear();
+
+                db.Households.Remove(household);
+                db.SaveChanges();
+            }
+            return RedirectToAction("Index", "Home");
         }
 
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
+        protected override void Dispose(bool disposing) {
+            if( disposing ) {
                 db.Dispose();
             }
             base.Dispose(disposing);
